+++ ProjectEulerSolutions/1-multiples-of-3-and-5.c
