+++ ProjectEulerSolutions/12-highly_divisible_triangular_number.c
@@ -1,0 +1,50 @@
+#include <stdio.h>
+#include <math.h>
+
+/**
+ * Project Euler Problem 12
+ *
+ * The sequence of triangle numbers is generated by adding the natural numbers.
+ * So the 7th triangle number would be 1 + 2 + 3 + 4 + 5 + 6 + 7 = 28.
+ * The first ten terms would be:
+ *
+ * 1, 3, 6, 10, 15, 21, 28, 36, 45, 55, ...
+ *
+ * Let us list the factors of the first seven triangle numbers:
+ *
+ *   1: 1
+ *   3: 1,3
+ *   6: 1,2,3,6
+ *  10: 1,2,5,10
+ *  15: 1,3,5,15
+ *  21: 1,3,7,21
+ *  28: 1,2,4,7,14,28
+ *
+ * We can see that 28 is the first triangle number to have over five divisors.
+ *
+ * What is the value of the first triangle number to have over five hundred divisors?
+ */
+
+int divisors(int triangle){
+    int count=0;
+    for (int i = (int)sqrt(triangle); i > 0; i--)
+    {
+        if (triangle % i == 0)
+        {
+            count++;
+        }
+    }
+    return 2*count;
+}
+
+int main() {
+    int row = 0, triangle = 0, count = 0, max = 500;
+
+    while(count < max){
+        triangle += row++;
+        count = divisors(triangle);
+    }
+
+    printf("%d has the required divisors greater than %d with %d total divisors\n", triangle, max, count);
+    return 0;
+}
